requirements.txt
# Core Dependencies
openai>=1.0.0
faiss-cpu>=1.7.4
gradio>=4.0.0
tiktoken>=0.5.0
python-docx>=1.0.0

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0
